// frontend/src/app/core/api/generated/model/verificationRuleEnum.ts
/**
 * Chemistry Apps API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `aromatic` - aromatic * `hbond_donor` - hbond_donor * `hbond_acceptor` - hbond_acceptor * `hydrophobic` - hydrophobic * `smarts` - smarts
 */
export const VerificationRuleEnum = {
    Aromatic: 'aromatic',
    HbondDonor: 'hbond_donor',
    HbondAcceptor: 'hbond_acceptor',
    Hydrophobic: 'hydrophobic',
    Smarts: 'smarts'
} as const;
export type VerificationRuleEnum = typeof VerificationRuleEnum[keyof typeof VerificationRuleEnum];